Actualización de productos, seeder y migraciones

+++ database/migrations/2025_04_29_212447_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->text('description')->nullable();
-            $table->decimal('price', 8, 2);
+            $table->integer('price');
             $table->string('image')->nullable();
             $table->timestamps();
         });
